CORE-3159: Logger class cast exception running command line. Check logging implementation before attempting to cast and configure console appender. Moved all logging implementation specific code into a single class.
