+++ liquibase-core/src/main/java/liquibase/integration/commandline/CommandLineOutputAppender.java
@@ -1,14 +1,24 @@
 package liquibase.integration.commandline;
 
+import ch.qos.logback.classic.Level;
+import ch.qos.logback.classic.Logger;
 import ch.qos.logback.classic.LoggerContext;
 import ch.qos.logback.classic.encoder.PatternLayoutEncoder;
+import ch.qos.logback.classic.spi.ILoggingEvent;
 import ch.qos.logback.classic.spi.LoggingEvent;
+import ch.qos.logback.core.Appender;
 import ch.qos.logback.core.ConsoleAppender;
+import ch.qos.logback.core.filter.AbstractMatcherFilter;
 import ch.qos.logback.core.filter.Filter;
 import ch.qos.logback.core.joran.spi.ConsoleTarget;
 import ch.qos.logback.core.spi.FilterReply;
+import liquibase.logging.LogLevel;
 import liquibase.logging.LogType;
+
+import java.util.Iterator;
+
 import org.slf4j.ILoggerFactory;
+import org.slf4j.LoggerFactory;
 import org.slf4j.Marker;
 
 /**
@@ -80,5 +90,44 @@ public class CommandLineOutputAppender extends ConsoleAppender {
          * Outputs all log messages to stdout
          */
         FULL_LOG,
+    }
+
+    /**
+     * Set up Logback logging to the STDOUT/STDERR console streams.
+     */
+    protected static void setupLogging(final Logger root, final LogLevel defaultLogLevel) {
+        // NOTE: Mismatched levels default to debug.
+        root.setLevel(Level.toLevel(defaultLogLevel.name()));
+    
+        ConsoleLogFilter consoleLogFilter = new ConsoleLogFilter();
+
+        Iterator<Appender<ILoggingEvent>> appenderIterator = root.iteratorForAppenders();
+        while (appenderIterator.hasNext()) {
+            Appender<ILoggingEvent> next = appenderIterator.next();
+            if (next instanceof ConsoleAppender) {
+                ((ConsoleAppender) next).addFilter(consoleLogFilter);
+            }
+        }
+    
+        for (String target : new String[] { "System.out", "System.err" }) {
+            CommandLineOutputAppender appender = new CommandLineOutputAppender(LoggerFactory.getILoggerFactory(), target);
+            root.addAppender(appender);
+            appender.start();
+        }
+    }
+
+    private static class ConsoleLogFilter extends AbstractMatcherFilter {
+
+        // NOTE: This is never set so the reply is always DENY.
+        private boolean outputLogs;
+
+        @Override
+        public FilterReply decide(Object event) {
+            if (outputLogs) {
+                return FilterReply.ACCEPT;
+            } else {
+                return FilterReply.DENY;
+            }
+        }
     }
 }
